Create page grabber script
It will attempt to fetch files based on a url path and increment until
it cannot fetch files. This is determined by the status code associated
with the attempt to fetch the file. This will only check for jpg and png
as they are the only file extensions I've seen for the particular CDN I
built this script for.